Add &nbsp; to empty table cells.
In a lot of browsers, entirely empty table cells are not rendered, leading to
lack of borders.

Or,more descriptively, from the HTML3.2 spec:

"Tables are commonly rendered in bas-relief, raised up with the outer border as
a bevel, and individual cells inset into this raised surface. Borders around
individual cells are only drawn if the cell has explicit content."

Having empty cells render the same as normal cells seems to be the normal user's
expectation, and having this not happen is considered as looking "broken".

These days it can be easily fixed using CSS:

    table { empty-cells: show; }

... but it's nicer if it Just Works.